fixed compatibility issue between bcrypt and passlib

+++ pawGPT-back-end/app/config.py
@@ -29,5 +29,5 @@ with open("certs/public_key.pem", "r") as f:
     PUBLIC_KEY = f.read()
 
 ALGORITHM = os.environ.get("ALGORITHM")
-ACCESS_TOKEN_EXPIRE_MINUTES: int = os.environ.get("ACCESS_TOKEN_EXPIRE_MINUTES")
-REFRESH_TOKEN_EXPIRE_DAYS: int = os.environ.get("REFRESH_TOKEN_EXPIRE_DAYS")
+ACCESS_TOKEN_EXPIRE_MINUTES: float = float(os.environ.get("ACCESS_TOKEN_EXPIRE_MINUTES"))
+REFRESH_TOKEN_EXPIRE_DAYS: float = float(os.environ.get("REFRESH_TOKEN_EXPIRE_DAYS"))
